refactor(command-palette): name components in full
CP.vue -> CommandPalette.vue, CPRow.vue -> CommandPaletteRow.vue.
The abbreviations were the only ones in the folder and read as noise
next to commands.ts / searchCommands.ts.
