+++ desk/src/components/command-palette/CommandPaletteRow.vue
